refactor: split worker into modules
